requirements.txt
Django==5.1.4
djangorestframework==3.16.0
gunicorn==22.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==7.4.0
pytest-django==4.5.2
